requirements.txt
easyocr==1.7.2
Faker==36.1.1
numpy==2.2.3
Pillow==11.1.0
protobuf==5.29.3
pyautogui==0.9.54
python-dotenv==1.0.1
pytube==15.0.0
selenium==4.29.0
google-generativeai==0.8.4